pyproject.toml
[project]
name = "access-permission-ml"
version = "0.1.0"
description = "Access permission approval prediction with CatBoost and user-grouped CV."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "yakvrz", email = "yakvrz@gmail.com" },
]
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "catboost>=1.2",
  "pyyaml>=6.0",
  "mlflow>=2.16",
  "great-expectations>=0.17,<0.18",
  "evidently>=0.4,<0.5",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "prefect>=2.14",
  "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=23.9",
  "ruff>=0.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
